AI-Assitant/requirements.txt
langdetect
tensorflow
docx
farm-haystack
farm-haystack[file-conversion]
farm-haystack[preprocessing]
